requirements.txt
# Copyright All rights Reserved 2025-2030, Ashutosh Sinha, Email: ajsinha@gmail.com
# Python dependencies for SAJHA MCP Server

# Core Framework
Flask>=3.0.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0
Flask-Session>=0.5.0

# WebSocket support
python-socketio>=5.9.0
eventlet>=0.33.0
gevent>=23.9.0

# JSON Schema Validation
jsonschema>=4.19.0

# Configuration & Utilities
python-dotenv>=1.0.0

# Monitoring & Metrics
prometheus-client>=0.18.0

# Logging
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0

# Additional utilities
requests>=2.31.0
urllib3>=2.0.0

# Databases
duckdb==1.4.1
psycopg2==2.9.11
PyMySQL==1.1.2
SQLAlchemy==2.0.31

# LLM Integration
anthropic>=0.39.0
langgraph>=0.2.0
langchain-core>=0.3.0


